fix the --illustrate option so that it scales the hashes correctly
The hashes were not scaled to correctly overlay the spectrogram; they
were squached and unreadable. I also changed it to display the hashes
from the reference where the match was found in addition to the hashes
from the query and those hashes that matched. There are also now
checkboxes for turning each of these three on or off.

Fixes #65.